chore(example): rename basic folder to js
