--- examples/basic/base.bash
@@ -1,4 +1,0 @@
-#!/bin/bash
-echo What is your name?
-read name
-echo Hello, $name!
